redmine_telegram_bot/requirements.txt
Django==3.2.18
python-telegram-bot==13.11
celery==5.2.6
redis==4.3.4
django-celery-beat==2.5.0
psycopg2-binary==2.9.5
gunicorn==20.0.4
django-cors-headers==3.14.0
python-dotenv

